LayoutTests/platform/ios-simulator/media/modern-media-controls/airplay-button/airplay-button-on-expected.txt

Testing the AirplayButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS airplayButton.element.classList.contains("on") is true
window.getComputedStyle(airplayButton.element).backgroundColor = rgb(0, 122, 255)
PASS successfullyParsed is true

TEST COMPLETE
